chore(release): cut v0.9.1
Popover polish + shared dismiss hook.

- Agents popover now shares its layout vocabulary with the Selector
  and History popovers: same `top` gap below the statusbar
  (`calc(100% + 2px)`, was `+ 6px`), same design tokens
  (`var(--z-popover)` / `var(--radius)` / `--vscode-sideBar-background`
  fallback), same `0 8px 24px / .32` drop shadow.
- Agents popover default width capped at 300px (was 360px) so chat
  content stays visible behind it on narrow side panels. Row content
  still fits with ellipsis on overflow.
- All three header popovers (Selector, Agents, History) share a new
  `useDismissableMenu` hook (`webview/src/hooks/useDismissableMenu.ts`):
  one source for open state, ref, outside-click + Escape listeners.
  Action handlers use the semantic `close()` helper instead of
  `setOpen(false)`.

Closes #154
