+++ package.json
@@ -3,7 +3,7 @@
   "name": "opencui",
   "displayName": "OpenCode Panel",
   "description": "OpenCode Panel: opencode-powered AI assistant with a modern React chat UI",
-  "version": "0.9.0",
+  "version": "0.9.1",
   "private": true,
   "publisher": "haoyangzeng",
   "license": "MIT",
